feat(console): user settings unsaved changes alert
